requirements.txt
fastapi
uvicorn
python-dotenv
chromadb
sentence-transformers
PyPDF2
pdfplumber
openai
requests
pydantic
anthropic
httpx
